==> CSGO Major players analysis/1_preprocessing_basic_player_info.ipynb <==
**LUIS IMLAUER | 2023 | limlauer.github.io**
# CS:GO Major players analysis

**WHO IS THE CS:GO GOAT?**
Majors are the biggest tournaments in CS:GO, and with the launch of CS2, we saw the last one forever. It's time to collect the data and analyze everything we can and answer some questions.

**FINAL DASHBOARD**
You can check the finished Dashboard here: https://public.tableau.com/app/profile/luis2877/viz/CSGO-Majors-Stats_16939594005680/CSGOMajorstats?publish=yes

## WHAT WE WILL EXPLORE
1. Who played the most majors?
2. Who played the most matches
3. Who played the most rounds
4. Who has the most kills
5. Who has the least kills
6. Who has the most deaths
7. Who are the majors MVPs and what role did they have
8. What country has the players with most majors
9. How did the 'online era' affect players and teams
10. How did the Ukraine war affect players
11. Winners of every major
12. Who has the most majors
13. Who performed better accross all majors (In total and in average)
14. Who is the CS:GO GOAT?


**IMPORTANT**
- In some statistics we will only be using info from players who attended 3+ majors
- To this date there is no hltv.org API available, so all information was gathered manually

**SOURCES**
- hltv.org
- https://www.kaggle.com/datasets/matheusnbrega/players-stats-all-csgo-majors

**LET'S START**
- I will be doing preprocessing and data gathering in python, and creating visualizations in Tableau.

import pandas as pd
pd.options.display.max_columns = 20
pd.options.display.max_rows = 2500


df = pd.read_csv('major-stats.csv', sep=';') # This file is separated by ;
df.head()

df['name'].value_counts().unique

df[df['name'] == 'dupreeh']

# I know dupreeh attended every major, but there were 19 of them, not 23
df[df['name'] == 'dupreeh']['name'].count()
# As we can see above, there is information about the "Challengers Stage"
# We don't need that, so we will just drop those rows

df = df[~df['event'].str.contains('Challengers Stage')]
df[df['name'] == 'dupreeh']['name'].count()

df['event'].unique()

### 1 - Who played the most majors?

# Perfect, now we have the general statistics for every player at every event.

print("Majors played")
for name, count in df['name'].value_counts().head(10).items():
    print(f"{name:13}>> {count} majors")

### 2 - Who played the most matches?

print("Total matches played")
df.groupby('name')['maps'].sum().sort_values(ascending=False).head(10)

### 3 - Who played the most rounds?

print("Total rounds played")
df.groupby('name')['rounds'].sum().sort_values(ascending=False).head(10)

## Adding more data

# We need more info that isn't included in the dataset. Let's do some scraping (no API yet for HLTV.org)
# First let's find the id for the players, the events, the URL format for each player in each event, and get the info from those pages.

# Dupreeh's link: https://www.hltv.org/player/7398/dupreeh
# https://www.hltv.org/stats/players/events/7398/dupreeh - all events played by dupreeh
# https://www.hltv.org/stats/players/individual/7398/dupreeh?event=6972 - individual stats at "BLAST Premier Spring Final 2023" (6972)

# Adding the event ID to each row
event_ids_dict = {
    'BLAST.tv Paris Major 2023': 6973, # https://www.hltv.org/events/6793/blasttv-paris-major-2023 For some reason, this is 6972 for dupreeh
    'IEM Rio Major 2022': 6586, # https://www.hltv.org/events/6586/iem-rio-major-2022
    'PGL Major Antwerp 2022': 6372, # https://www.hltv.org/events/6372/pgl-major-antwerp-2022
    'PGL Major Stockholm 2021': 4866, # https://www.hltv.org/events/4866/pgl-major-stockholm-2021
    'StarLadder Major Berlin 2019': 4443, # https://www.hltv.org/events/4443/starladder-major-berlin-2019
    'IEM Katowice 2019': 3883, # https://www.hltv.org/events/3883/iem-katowice-2019
    'FACEIT Major 2018': 3564, # https://www.hltv.org/events/3564/faceit-major-2018
    'ELEAGUE Major 2018': 3247, # https://www.hltv.org/events/3247/eleague-major-2018
    'PGL Major Krakow 2017': 2720, # https://www.hltv.org/events/2720/pgl-major-krakow-2017
    'ELEAGUE Major 2017': 2471, # https://www.hltv.org/events/2471/eleague-major-2017
    'ESL One Cologne 2016': 2062, # https://www.hltv.org/events/2062/esl-one-cologne-2016
    'MLG Columbus 2016': 2027, # https://www.hltv.org/events/2027/mlg-columbus-2016
    'DreamHack Open Cluj-Napoca 2015': 1617, # https://www.hltv.org/events/1617/dreamhack-open-cluj-napoca-2015
    'ESL One Cologne 2015': 1666, # https://www.hltv.org/events/1666/esl-one-cologne-2015
    'ESL One Katowice 2015': 1611, # https://www.hltv.org/events/1611/esl-one-katowice-2015
    'DreamHack Winter 2014': 1553, # https://www.hltv.org/events/1553/dreamhack-winter-2014
    'ESL One Cologne 2014': 1444, # https://www.hltv.org/events/1444/esl-one-cologne-2014
    'EMS One Katowice 2014': 1333, # https://www.hltv.org/events/1333/ems-one-katowice-2014
    'DreamHack Winter 2013': 1270 # https://www.hltv.org/events/1270/dreamhack-winter-2013
}

df['event_ID'] = df['event'].map(event_ids_dict)
df.loc[(df['name'] == 'dupreeh') & (df['event'] == 'BLAST.tv Paris Major 2023'), 'event_ID'] = 6972
df.info()

# Import libraries
# Had to use selenium
from selenium import webdriver
driver = webdriver.Chrome()
from bs4 import BeautifulSoup
import requests
import smtplib

# Let's test if it works first

URL = "https://www.hltv.org/stats/players/individual/7398/dupreeh?event=6972"
HEADERS = {"User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/115.0.0.0 Safari/537.36", "Accept-Encoding":"gzip, deflate", "Accept":"text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8", "DNT":"1","Connection":"close", "Upgrade-Insecure-Requests":"1"}

driver = webdriver.Chrome()
driver.get(URL)
page_source = driver.page_source
driver.quit()

soup1 = BeautifulSoup(page_source, "html.parser")
soup2 = BeautifulSoup(soup1.prettify(), "html.parser")

stats_rows = soup2.find_all('div', class_='stats-row')
for row in stats_rows:
    stat_name = row.find('span').text.strip()
    stat_value = row.find_all('span')[1].text.strip()
    print(f"{stat_name}: {stat_value}")


It works! We can now add all of that data to our dataset, it will take some time but we should only be doing this once for every row.

df['player_id'] = 0
df.loc[df['name'] == 'dupreeh', 'player_id'] = 7398

# Let's scrape some more and after we are finished, check if the data is complete and good to go
# There are 1521 rows in total.
# Let's start with dupreeh

player_id = 7398
player_name = 'dupreeh'
event_event_ID = 6972
URL = f"https://www.hltv.org/stats/players/individual/{player_id}/{player_name}?event={event_event_ID}"
#HEADERS = {"User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/115.0.0.0 Safari/537.36", "Accept-Encoding":"gzip, deflate", "Accept":"text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8", "DNT":"1","Connection":"close", "Upgrade-Insecure-Requests":"1"}

driver = webdriver.Chrome()
driver.get(URL)
page_source = driver.page_source
driver.quit()

soup1 = BeautifulSoup(page_source, "html.parser")
soup2 = BeautifulSoup(soup1.prettify(), "html.parser")

stats_rows = soup2.find_all('div', class_='stats-row')
for row in stats_rows:
    stat_name = row.find('span').text.strip()
    stat_value = row.find_all('span')[1].text.strip()
    print(f"{stat_name}: {stat_value}")

There we printed every piece of data we can gather from that page. Let's divide that into columns and add it gradually to the df. Note that KD difference shows as K - D diff. so we should fix that before adding it to the database

But first, we need to get every player id on hltv

import re

df['name'].unique().__len__()

# 391 different players, that's too much to do manually
# Let's try to automate this task using the search bar on hltv.org

# This script below took me hours to code and test, and I should have used regex for everything (I think); but as soon as it worked I just wanted to keep going, I'll maybe fix it in the future
def extract_player_info(csv_name, Player_Names_List, Overwrite, Start_index = 0, Ending_index = 0):
    try:
        loaded_df = pd.read_csv(csv_name)
        print("Loaded existing CSV.")
    except FileNotFoundError:
        print("CSV file not found. Creating a new DataFrame.")
        loaded_df = pd.DataFrame(columns=['Name', 'Full name', 'Player ID', 'Full link', 'Multiple'])
    Output_DF = loaded_df

    if (Ending_index == 0):
        Ending_index = Player_Names_List.unique().__len__()

    print(f"Preparing to search for {Ending_index} players...")
    player_names = Player_Names_List.unique()[Start_index:Ending_index]
    multiple_results, p_fullname, p_playerid, p_fullLink = False, 'Error', 'Error', 'Error'

    names = list(Output_DF['Name'])

    for p_name in player_names:
        if (p_name in names) and (not Overwrite):
            print(f"{p_name} already searched, skipping")
        else:
            multiple_results = False
            URL = f"https://www.hltv.org/search?query={p_name.lower()}"
            driver = webdriver.Chrome()
            driver.get(URL)
            page_source = driver.page_source
            driver.quit()
            soup1 = BeautifulSoup(page_source, "html.parser")

            tables = soup1.find_all('table', class_='table')
            starting, ending = 0, 0 # starting and ending indexes
            for i, row in enumerate(str(tables).split('\n')):
                if row == '<td class="table-header">Player</td>':
                    starting = i + 3
                if row == '<td class="table-header">Article</td>':
                    ending = i - 3
            results = str(tables).split('\n')[starting:ending]
            
            if len(results) > 1:
                multiple_results = True
            
            results = results[0]
            p_fullname = (results.split('"/>')[1]).split('</a>')[0]
            p_playerid = results.split('/')[2]
            pattern = r'href="(/player/\d+/[^"]+)"'
            match = re.search(pattern, results)
            if match:
                p_fullLink = match.group(1)

            print(f"{p_name} -> {multiple_results}, {p_fullname}, {p_playerid}, https://www.hltv.org{p_fullLink}")
            
            new_player_dict = {
                'Name': p_name,
                'Full name': p_fullname, 
                'Player ID': p_playerid, 
                'Full link': 'https://www.hltv.org' + p_fullLink, 
                'Multiple': multiple_results
            }
            
            Output_DF = pd.concat([Output_DF, pd.DataFrame([new_player_dict])], ignore_index=True)

    Output_DF.to_csv("player_hltv_info.csv", index=False)
    print("File saved as player_hltv_info.csv")


extract_player_info("player_hltv_info.csv", df['name'], False, Ending_index = 25)

It works! Now let's do it for every player (~11.5 seconds per player, it should take about 1:10 hours)

extract_player_info("player_hltv_info.csv", df['name'], False)

try:
        loaded_df = pd.read_csv("player_hltv_info.csv")
        print("Loaded existing CSV.")
except FileNotFoundError:
    print("CSV file not found. Creating a new DataFrame.")

loaded_df.head()

Some searches return multiple results (For example: iM returns "s1mple" as the first result), so let's see which players are imported correctly and which aren't.

1. Filter by Multiple
2. Check every player

percentage_true = (loaded_df[loaded_df['Multiple'] == True].shape[0] / loaded_df.shape[0]) * 100
print(f"Percentage of 'Multiple' = True: {percentage_true:.0f}%")

To be honest, I expected this result. Some players will be easy to classify as correct/incorrect.
I will do that in excel and re-import the filtered file.

• IN EXCEL:
1. I filtered by "Multiple = True" and added the "Checked" and "Correct" column
2. Then, I fixed every link that was wrong. This was because player names were similar, or because some of them are now coaches.
3. Finally, I used a Search function to find the name of the player in the link, and see if it was referring to that player.
4. After I manually checked everything I saved it and we are ready to continue our analysis.

try:
        loaded_df = pd.read_csv("player_hltv_info_check.csv")
        print("Loaded existing CSV.")
except FileNotFoundError:
    print("CSV file not found. Creating a new DataFrame.")

loaded_df.head()

There are some enconding errors, we could fix those but we can also just fix the ones we see in the final results, which I think is easier.

We gathered great data, but now it's time to merge it with our original df and save everything!

merged_df = df.merge(loaded_df, left_on='name', right_on='Name', how='left')
merged_df.head()

# Let's drop the second 'Name' column
merged_df = merged_df.drop(['Name'], axis = 1)
merged_df.head()

# And that 'player_id' column too
merged_df = merged_df.drop(['player_id'], axis = 1)
merged_df.head()

Great, now let's add more information, specifically the individual stats from a player in a given event.
(The example data is what we pulled from dupreeh in the last major)

We will be adding:
1. OVERALL STATS
- Kills: 200
- Deaths: 233
- Kill / Death: 0.86
- Kill / Round: 0.61
- Rounds with kills: 138
- Kill - Death difference: K - D diff.

2. OPENING STATS
- Total opening kills: 31
- Total opening deaths: 44
- Opening kill ratio: 0.70
- Opening kill rating: 0.91
- Team win percent after first kill: 80.6%
- First kill in won rounds: 15.2%

3. ROUND STATS
- 0 kill rounds: 192
- 1 kill rounds: 95
- 2 kill rounds: 26
- 3 kill rounds: 15
- 4 kill rounds: 2
- 5 kill rounds: 0

4. WEAPON STATS
- Rifle kills: 153
- Sniper kills: 1
- SMG kills: 15
- Pistol kills: 31
- Grenade: 0
- Other: 1

# Continues in the next notebook!
merged_df.to_csv("basic_players_info.csv", index=False)
print("File saved as basic_players_info.csv")
