==> CSGO Major players analysis/3_completing_individual_stats.ipynb <==
import pandas as pd
pd.options.display.max_columns = 20
pd.options.display.max_rows = 2500

df = pd.read_csv('basic_players_info.csv') # Load our extracted info
df2 = pd.read_csv('player_individual_stats.csv') # Load our extracted info
df.head()

df2.head()


df3 = pd.concat([df, df2], axis=1)
df3.head()

df3.to_csv('salida.csv', index = False)

There are a lot of missing values, let's fix that

df3.tail(10)

Missing values come in the form of NaN (not a number) and 0 (the player ID and event ID combination didn't exist on hltv so it displayed 0).

First let's fix the NaNs

df_nans = df3[df3['3 kill rounds'].isna()]
df_nans.info()

"Index: 110 entries, 1411 to 1520"

Those are the rows we need to fix

from selenium import webdriver
driver = webdriver.Chrome()
from bs4 import BeautifulSoup
driver.quit()

df_nans.columns = ['name', 'nationality', 'team', 'maps', 'rounds', 'KD-diff', 'KD',
       'rating', 'event', 'event_ID', 'Full name', 'Player ID', 'Full link',
       'name2', 'Player ID2', 'event_ID2', 'Kills', 'Deaths', 'Kill / Death',
       'Kill / Round', 'Rounds with kills', 'Kill - Death difference',
       'Total opening kills', 'Total opening deaths', 'Opening kill ratio',
       'Opening kill rating', 'Team win percent after first kill',
       'First kill in won rounds', '0 kill rounds', '1 kill rounds',
       '2 kill rounds', '3 kill rounds', '4 kill rounds', '5 kill rounds',
       'Rifle kills', 'Sniper kills', 'SMG kills', 'Pistol kills', 'Grenade',
       'Other', 'Event link']

Output_DF = pd.DataFrame(columns = df_nans.columns)

processing_list = df_nans
for index, row in processing_list.iterrows():
    Ending_index = 122
    total_rows = 122
    save_index = 0
    File_save_iter = 10

    player_id = row['Player ID']
    event_id = row['event_ID']
    player_name = row['name']

    URL = f"https://www.hltv.org/stats/players/individual/{player_id}/{player_name.lower()}?event={event_id}"
    driver = webdriver.Chrome()
    driver.get(URL)
    page_source = driver.page_source
    driver.quit()

    soup1 = BeautifulSoup(page_source, "html.parser")
    soup2 = BeautifulSoup(soup1.prettify(), "html.parser")

    stats_rows = soup2.find_all('div', class_='stats-row')
    stat_dict = {}  # Create a new empty dictionary
    print(f"[{index}/{Ending_index} | {total_rows}] Extracting data for {player_name}...")
    stat_dict['name2'] = player_name
    stat_dict['Player ID2'] = player_id
    stat_dict['event_ID2'] = event_id
    stat_dict['Event link'] = f"https://www.hltv.org/stats/players/individual/{player_id}/{player_name}?event={event_id}"
    for row in stats_rows:
        stat_name = row.find('span').text.strip()
        stat_value = row.find_all('span')[1].text.strip()
        stat_dict[stat_name] = stat_value  # Add to dictionary
    
    stat_dict['Kill - Death difference'] = int(stat_dict['Kills']) - int(stat_dict['Deaths'])

    Output_DF = pd.concat([Output_DF, pd.DataFrame([stat_dict])], ignore_index=True)
    print(f"[{index}/{Ending_index} | {total_rows}] Added statistics for Player {player_name} with ID {player_id} in event {event_id}")
    save_index += 1
    if save_index >= File_save_iter:
        Output_DF.to_csv("nans_stats.csv", index=False)
        print(f"[SAVED] > File saved after {File_save_iter} iterations, as nans_stats.csv")
        save_index = 0
    
Output_DF.to_csv("nans_stats.csv", index=False)
print("[SAVED] > File saved as nans_stats.csv")
