==> CSGO Major players analysis/4_completing_missing_data.ipynb <==
import pandas as pd
from selenium import webdriver
driver = webdriver.Chrome()
from bs4 import BeautifulSoup
pd.options.display.max_columns = 20
pd.options.display.max_rows = 2500

df = pd.read_csv('player_individual_stats_complete.csv') # This is the file we saved in the previous notebook
df.head()

df.info()

print(df[df['0 kill rounds'] == 0].count())

df_missing = df[df['0 kill rounds'] == 0]
df_missing.head()
#df_missing.to_csv("individuals_missing_values.csv", index=False)

df_mistakes = pd.read_csv("6793_fixed_players.csv")

df_mistakes[df_mistakes['event_ID'] == 6793].count()

try:
    loaded_df = pd.read_csv("6793_tofix_players.csv")
    print("Loaded existing CSV.")
except FileNotFoundError:
    print("CSV file not found.")
    exit

Output_DF = loaded_df

print(f"Preparing to search for {Output_DF[Output_DF['event_ID'] == 6793].count()} players...")
processing_list = Output_DF

existing_player_ids = list(Output_DF['Player ID']) # EXISTENTES
existing_event_ids = list(Output_DF['event_ID']) # EXISTENTES

for index, row in processing_list.iterrows():
    player_id = row['Player ID']
    event_id = row['event_ID']
    player_name = row['name']

    if event_id == 6793:
        URL = f"https://www.hltv.org/stats/players/individual/{player_id}/{player_name.lower()}?event=6793"
        driver = webdriver.Chrome()
        driver.get(URL)
        page_source = driver.page_source
        driver.quit()

        soup1 = BeautifulSoup(page_source, "html.parser")
        soup2 = BeautifulSoup(soup1.prettify(), "html.parser")

        stats_rows = soup2.find_all('div', class_='stats-row')
        stat_dict = {}  # Create a new empty dictionary
        print(f"[{index}] Extracting data for {player_name}...")
        stat_dict['name'] = player_name
        stat_dict['Player ID'] = player_id
        stat_dict['event_ID'] = 6793
        stat_dict['Event link'] = f"https://www.hltv.org/stats/players/individual/{player_id}/{player_name}?event=6793"
        for row in stats_rows:
            stat_name = row.find('span').text.strip()
            stat_value = row.find_all('span')[1].text.strip()
            stat_dict[stat_name] = stat_value  # Add to dictionary
        
        stat_dict['Kill - Death difference'] = int(stat_dict['Kills']) - int(stat_dict['Deaths'])

        Output_DF = pd.concat([Output_DF, pd.DataFrame([stat_dict])], ignore_index=True)
        print(f"[{index}] Added statistics for Player {player_name} with ID {player_id} in event {event_id}")
    
Output_DF.to_csv("6793_players_fixed.csv", index=False)
print("[SAVED] > File saved as 6793_players_fixed.csv")
