==> CSGO Major players analysis/5_checking_the_data.ipynb <==
We finally have all our data, I had to clean some of it in Excel, and mix every csv, but it worked.

import pandas as pd

df = pd.read_csv('6793_players_fixed.csv') # This is the file we saved in the previous notebook
df.head()

df.nunique()

duplicated_names_with_different_ids = df.groupby('name')['Player ID'].nunique()
duplicated_names_with_different_ids = duplicated_names_with_different_ids[duplicated_names_with_different_ids > 1]
print(duplicated_names_with_different_ids)

duplicated_names_with_different_ids = df.groupby('idevent')['Player ID'].nunique()
duplicated_names_with_different_ids = duplicated_names_with_different_ids[duplicated_names_with_different_ids > 1]
print(duplicated_names_with_different_ids)

# Perfect, we got our dataset, now let's analyze.

We can now continue with our analysis!

### 4. Who has the most kills

# Agrupar por 'Player ID' y sumar las kills
player_kills = df.groupby('Player ID')['Kills'].sum()

# Encontrar el Player ID con la mayor cantidad de kills
max_kills_player_id = player_kills.idxmax()

# Obtener el nombre y la cantidad de kills del jugador con más kills
max_kills_name = df[df['Player ID'] == max_kills_player_id]['name'].iloc[0]
max_kills_total = player_kills[max_kills_player_id]

print(f"Player with most kills: {max_kills_name}")
print(f"Player ID: {max_kills_player_id}")
print(f"Total kills: {max_kills_total}")

### 5. Who has the least kills

# Agrupar por 'Player ID' y sumar las kills
player_kills = df.groupby('Player ID')['Kills'].sum()

# Encontrar el Player ID con la mayor cantidad de kills
max_kills_player_id = player_kills.idxmin()

# Obtener el nombre y la cantidad de kills del jugador con más kills
max_kills_name = df[df['Player ID'] == max_kills_player_id]['name'].iloc[0]
max_kills_total = player_kills[max_kills_player_id]

print(f"Player with most kills: {max_kills_name}")
print(f"Player ID: {max_kills_player_id}")
print(f"Total kills: {max_kills_total}")

### 6. Who has the most deaths

# Agrupar por 'Player ID' y sumar las kills
player_deaths = df.groupby('Player ID')['Deaths'].sum()

# Encontrar el Player ID con la mayor cantidad de kills
max_deaths_player_id = player_deaths.idxmax()

# Obtener el nombre y la cantidad de kills del jugador con más kills
max_deaths_name = df[df['Player ID'] == max_deaths_player_id]['name'].iloc[0]
max_deaths_total = player_deaths[max_deaths_player_id]

print(f"Player with most deaths: {max_deaths_name}")
print(f"Player ID: {max_deaths_player_id}")
print(f"Total deaths: {max_deaths_total}")

### 7. Who are the majors MVPs and what role did they have

1. JW — MVP DreamHack Winter 2013 -> AWPer
2. PashaBiceps — MVP EMS One Katowice 2014 -> AWPer
3. Friberg — MVP ESL One Cologne 2014 -> Rifler
4. Happy — MVP DreamHack Winter 2014 -> Rifler
5. Olofmeister — MVP ESL One Katowice 2015 -> Rifler
6. Flusha — MVP ESL One Cologne 2015 -> Rifler
7. KennyS — MVP DreamHack Cluj-Napoca -> AWPer
8. Coldzera — MVP MLG Columbus 2016 & ESL One Cologne 2016 -> AWPer
9. Kjaerbye — MVP ELEAGUE Major Atlanta 2017 -> Rifler
10. AdreN — PGL Major Krakow 2017 MVP -> IGL / Rifler
11. Tarik — ELEAGUE Major Boston 2018 MVP -> Rifler
12. device — FaceIT Major London 2018 & StarLadder Major Berlin 2019 MVP -> AWPer
13. Magisk — IEM Katowice 2019 MVP -> Rifler
14. s1mple — PGL Major Stockholm 2021 MVP -> AWPer
15. Rain — PGL Major Antwerp 2022 MVP -> Rifler
16. Jame — IEM Rio 2022 MVP -> AWPer
17. ZywOo — BLAST.tv Paris Major 2023 MVP -> AWPer

Source: https://blog.cs.money/en/esports/all-major-mvp-winners-in-csgo-history/

### 8. What country has the players with most majors

### 9. How did the 'online era' affect players and teams

### 10. How did the Ukraine war affect players

### 11. Winners of every major

### 12. Who has the most majors

### 13. Who performed better accross all majors (In total and in average)

### 14. Who is the CS:GO GOAT?
