==> CSGO Major players analysis/6_mixing_the_csvs.ipynb <==
salida.csv and final_individual_stats.csv

import pandas as pd

df2 = pd.read_csv('salida.csv') 
df = pd.read_csv('final_individual_stats.csv') 
print(df.head())
print(df2.head())

pd.set_option('display.max_rows', None)
sorted(df['event_ID'].unique()) == sorted(df2['event_ID'].unique())

print(sorted(df['event_ID'].unique()))
print(sorted(df2['event_ID'].unique()))

print(df.columns)
print(df2.columns)

merged_df = df.merge(df2, on=['event_ID', 'Player ID'], how='left')

merged_df.to_csv('merged_individual_stats.csv', index=False)

After this, everything I had to delete and fix I did in excel.
